requirements.txt
beautifulsoup4==4.12.2
blinker==1.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
command-not-found==0.3
cryptography==3.4.8
cssselect==1.2.0
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.3
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
jieba3k==0.35.1
joblib==1.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.2
more-itertools==8.10.0
netifaces==0.11.0
newspaper3k==0.2.8
nltk==3.8.1
oauthlib==3.2.0
Pillow==9.5.0
platformdirs==3.10.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
SecretStorage==3.3.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
systemd-python==234
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
wadllib==1.3.6
zipp==1.0.0
